longrun-risk-scanner/requirements.txt
# longrun-risk-scanner — Python 3.10+
PyYAML>=6.0.1
python-dotenv>=1.0.0
httpx>=0.26.0
pytest>=7.4.0
